# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Restaurant Reservation System",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/reservation-system/restaurants": {
            "get": {
                "description": "Retrieves multiple restaurants info from restaurants table in PostgreSQL",
                "tags": [
                    "restaurant"
                ],
                "summary": "Fetches restaurants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Number of restaurants to fetch",
                        "name": "limit",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Number of restaurants to omit",
                        "name": "offset",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant.Restaurants"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while fetching restaurants",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts new restaurant info to restaurants table in PostgreSQL",
                "tags": [
                    "restaurant"
                ],
                "summary": "Creates a restaurant",
                "parameters": [
                    {
                        "description": "New data",
                        "name": "new_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restaurant.RestaurantDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant.ID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while creating restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/restaurants/{restaurant_id}": {
            "get": {
                "description": "Retrieves restaurant info from restaurants table in PostgreSQL",
                "tags": [
                    "restaurant"
                ],
                "summary": "Gets a restaurant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant.RestaurantInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid restaurant ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while getting restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates restaurant info in restaurants table in PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurant"
                ],
                "summary": "Updates a restaurant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New info",
                        "name": "new_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restaurant.RestaurantInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid restaurant ID or data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while updating restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes restaurant info from restaurants table in PostgreSQL",
                "tags": [
                    "restaurant"
                ],
                "summary": "Deletes a restaurant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid restaurant ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while deleting restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "restaurant.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "restaurant.RestaurantDetails": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "restaurant.RestaurantInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "restaurant.Restaurants": {
            "type": "object",
            "properties": {
                "restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurant.RestaurantInfo"
                    }
                }
            }
        },
...
    }
}